Separate Point and Vector definitions into modules
