+++ src/lib.rs
@@ -1,4 +1,5 @@
 #[cfg(test)]
 mod tests;
 
-mod util;
+mod point;
+mod vector;
